refactor: activeTabs.js
reviewd: Emmanuel Oliveira
