+++ scripts/modules/activeTabs.js
@@ -1,20 +1,28 @@
-export default function activeTabs() {
-  const tabMenu = document.querySelectorAll("[data-tab='menu'] li")
-  const tabContent = document.querySelectorAll("[data-tab='content'] section")
-  tabContent[0].classList.add("active")
-
-  function activeTab(index) {
-    tabContent.forEach((section) => {
-      section.classList.remove("active")
-    })
-    const direcao = tabContent[index].dataset.anime
-    tabContent[index].classList.add("active", direcao)
+export default class TabNav {
+  constructor(menu, content) {
+    this.tabMenu = document.querySelectorAll(menu)
+    this.tabContent = document.querySelectorAll(content)
+    this.activeClass = "active"
   }
-  if (tabMenu.length && tabContent.length) {
-    tabMenu.forEach((itemMenu, index) => {
-      itemMenu.addEventListener("click", () => {
-        activeTab(index)
-      })
+  activeTab(index) {
+    this.tabContent.forEach((section) => {
+      section.classList.remove(this.activeClass)
     })
+    const direcao = this.tabContent[index].dataset.anime
+    this.tabContent[index].classList.add(this.activeClass, direcao)
+  }
+
+  // Adiciona os eventos as Tabs
+  addTabNavEvent() {
+    this.tabMenu.forEach((itemMenu, index) => {
+      itemMenu.addEventListener("click", () => this.activeTab(index))
+    })
+  }
+  init() {
+    if (this.tabMenu.length && this.tabContent.length) {
+      // Ativar primeiro item
+      this.activeTab(0)
+      this.addTabNavEvent()
+    }
   }
 }
